pyproject.toml
[tool.poetry]
name = "nautobot_ansible_modules"
version = "5.6.0"
description = "Ansible collection to interact with Nautobot's API"
authors = ["Network to Code <opensource@networktocode.com>"]
license = "Apache 2.0"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
netutils = "^1.2"
# pynautobot = "^2.4.2"
ansible-core = "^2.16"
prance = "^25.4.8.0"
ruff = "^0.13.0"

[tool.poetry.group.dev.dependencies]
black = "*"
cryptography = "*"
jinja2 = "*"
jmespath = "*"
jsondiff = "*"
pytest = "*"
pytest-mock = "*"
pytest-xdist = "*"
pyyaml = "*"
mock = "*"
importlib-metadata = "*"
pylint = "*"
hypothesis = "*"
pytest-pythonpath = "*"
parameterized = "*"
invoke = "*"
bandit = "*"
pytest-forked = "*"
# Rendering docs to HTML
mkdocs = "1.6.1"
# Material for MkDocs theme
mkdocs-material = "9.5.30"
# Render custom markdown for version added/changed/remove notes
markdown-version-annotations = "1.0.1"
# Render dynamic pages for Ansible Collections documentation
mkdocs-ansible-collection = "0.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
python_paths = "./"
addopts = "-vv"

[tool.black]
line-length = 160
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long
    """

[tool.ruff]
line-length = 160
target-version = "py39"

[tool.ruff.lint]
ignore = [
    "E402",  # pycodestyle: module-import-not-at-top-of-file
]
